[REFACTOR] SlackErrorLogger 분리 #443
