+++ src/main/java/FIS/iLUVit/aspect/trace/filter/TraceFilter.java
@@ -1,6 +1,7 @@
 package FIS.iLUVit.aspect.trace.filter;
 
 import FIS.iLUVit.aspect.trace.TraceSupports;
+import FIS.iLUVit.exception.exceptionHandler.SlackErrorLogger;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.annotation.Order;
@@ -16,6 +17,8 @@ import java.io.IOException;
 @Slf4j
 public class TraceFilter implements Filter {
 
+    private final SlackErrorLogger slackErrorLogger;
+
     @Override
     public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
         HttpServletRequest request = (HttpServletRequest) servletRequest;
@@ -24,11 +27,12 @@ public class TraceFilter implements Filter {
             filterChain.doFilter(servletRequest, servletResponse);
             TraceSupports.releaseTraceInfos();
         }catch(Exception e){
-            log.error("[INTERNAL SERVER ERROR] {} {} errMessage={}\n",
+            log.error("[InternalServerError] {} {} errMessage={}\n",
                     request.getMethod(),
                     request.getRequestURI(),
                     e.getMessage()
             );
+            slackErrorLogger.sendSlackAlertErrorLog(e.getMessage(), request); // 슬랙 알림 보내는 메서드
             throw e;
         }
